refactor: fixed for the issue (#511)
