+++ packages/atproto/lib/atproto.dart
@@ -27,7 +27,6 @@ export 'package:atproto/src/entities/parsed_did_doc.dart';
 export 'package:atproto/src/entities/parsed_service.dart';
 export 'package:atproto/src/entities/parsed_verification_method.dart';
 export 'package:atproto/src/entities/record.dart';
-export 'package:atproto/src/entities/record_value.dart';
 export 'package:atproto/src/entities/records.dart';
 export 'package:atproto/src/entities/repo.dart';
 export 'package:atproto/src/entities/repo_block.dart';
