+++ packages/atproto/lib/src/entities/record.dart
@@ -2,10 +2,12 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
-// 📦 Package imports:
+// ignore_for_file: invalid_annotation_target
+
 import 'package:atproto_core/atproto_core.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+// 📦 Package imports:
 import 'strong_ref.dart';
 
 part 'record.freezed.dart';
@@ -16,13 +18,32 @@ class Record with _$Record {
   // ignore: unused_element
   const Record._();
 
+  @JsonSerializable(includeIfNull: false)
   const factory Record({
-    required String cid,
     @AtUriConverter() required AtUri uri,
+    String? cid,
+    required Map<String, dynamic> value,
   }) = _Record;
 
   factory Record.fromJson(Map<String, Object?> json) => _$RecordFromJson(json);
 
+  /// Returns true if this record has strong ref, otherwise false.
+  bool get hasStrongRef => cid != null;
+
+  /// Returns true if this record has not strong ref, otherwise false.
+  bool get hasNotStrongRef => !hasStrongRef;
+
   /// Returns the [StrongRef] representation of this record.
-  StrongRef toStrongRef() => StrongRef(cid: cid, uri: uri);
+  ///
+  /// Throws [StateError] if this record has not strong ref.
+  StrongRef toStrongRef() {
+    if (hasNotStrongRef) {
+      throw StateError(
+        'This record has not strong ref. '
+        'Check with "hasStrongRef" or "hasNotStrongRef."',
+      );
+    }
+
+    return StrongRef(cid: cid!, uri: uri);
+  }
 }
