+++ packages/atproto/pubspec.yaml
@@ -1,6 +1,6 @@
 name: atproto
 description: The easiest and powerful Dart/Flutter library for AT Protocol.
-version: 0.3.5
+version: 0.4.0
 repository: https://github.com/myConsciousness/atproto.dart/tree/main/packages/atproto
 issue_tracker: https://github.com/myConsciousness/atproto.dart/issues
 funding:
