+++ packages/bluesky/pubspec.yaml
@@ -1,6 +1,6 @@
 name: bluesky
 description: The easiest and powerful Dart/Flutter library for Bluesky Social.
-version: 0.5.8
+version: 0.6.0
 repository: https://github.com/myConsciousness/atproto.dart/tree/main/packages/bluesky
 issue_tracker: https://github.com/myConsciousness/atproto.dart/issues
 funding:
@@ -10,7 +10,7 @@ environment:
 
 dependencies:
   atproto_core: ^0.3.5
-  atproto: ^0.3.5
+  atproto: ^0.4.0
   freezed_annotation: ^2.2.0
   json_annotation: ^4.8.0
