+++ packages/bluesky/test/src/actors/actors_service_test.dart
@@ -393,7 +393,7 @@ void main() {
       );
 
       expect(response, isA<XRPCResponse>());
-      expect(response.data, isA<Record>());
+      expect(response.data, isA<StrongRef>());
     });
 
     test('when unauthorized', () async {
